requirements.txt
acme==1.1.0
asgiref==3.5.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certbot==0.40.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
colorama==0.4.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.4.0
Django==3.2.5
django-crispy-forms==1.14.0
django-filter==21.1
django-import-export==2.8.0
django-widget-tweaks==1.4.12
entrypoints==0.3
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jeepney==0.8.0
Jinja2==2.10.1
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupPy==1.14
MarkupSafe==1.1.0
mock==3.0.5
more-itertools==4.2.0
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.9
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py-zabbix==1.1.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
schedule==1.1.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
ssh-import-id==5.10
tablib==3.2.1
testresources==2.0.1
typing_extensions==4.6.3
urllib3==1.25.8
wadllib==1.3.3
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
